refactor utils directory
